+++ ParsMarkt/Interfaces/IPersonServices.cs
@@ -8,12 +8,12 @@ namespace ParsMarkt
     public interface IPersonServices
     {
         Task<IList<PersonViewModel>> GetAsync();
-        Task<PersonViewModel> GetByIdAsync(int id);
+        Task<PersonViewModel> GetByIdAsync(long id);
         Task<PersonViewModel> PostAsync(PersonViewModel viewModel);
         Task<PersonViewModel> PutAsync(PersonViewModel viewModel);
 
         Task<bool> DeleteAsync(int id);
         Task<PersonViewModel> PostPersonAsync(PersonViewModel viewModel);
-        Task<PersonViewModel> PutPersonAsync<PersonViewModel>(PersonViewModel viewModel, int id);
+        Task<PersonViewModel> PutPersonAsync<PersonViewModel>(PersonViewModel viewModel, long id);
     }
 }
